+++ boilerplate/backend/go-http/config/thirdpartypasswordless.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"github.com/supertokens/supertokens-golang/recipe/dashboard"
-	"github.com/supertokens/supertokens-golang/recipe/dashboard/dashboardmodels"
 	"github.com/supertokens/supertokens-golang/recipe/passwordless/plessmodels"
 	"github.com/supertokens/supertokens-golang/recipe/session"
 	"github.com/supertokens/supertokens-golang/recipe/thirdparty"
@@ -27,7 +26,7 @@ var SuperTokensConfig = supertokens.TypeInput{
 			ContactMethodEmailOrPhone: plessmodels.ContactMethodEmailOrPhoneConfig{
 				Enabled: true,
 			},
-			Providers: [] tpmodels.TypeProvider{
+			Providers: []tpmodels.TypeProvider{
 				thirdparty.Google(tpmodels.GoogleConfig{
 					ClientID:     "1060725074195-kmeum4crr01uirfl2op9kd5acmi9jutn.apps.googleusercontent.com",
 					ClientSecret: "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",
@@ -47,8 +46,6 @@ var SuperTokensConfig = supertokens.TypeInput{
 			},
 		}),
 		session.Init(nil), // initializes session features
-		dashboard.Init(dashboardmodels.TypeInput{
-			ApiKey: "supertokens_is_awesome",
-		}),
+		dashboard.Init(nil),
 	},
 }