+++ boilerplate/backend/nest/config/thirdparty.ts
@@ -41,7 +41,5 @@ export const recipeList = [
     },
   }),
   Session.init(),
-  Dashboard.init({
-    apiKey: 'supertokens_is_awesome',
-  }),
+  Dashboard.init(),
 ];