+++ boilerplate/backend/nest/config/thirdpartyemailpassword.ts
@@ -39,7 +39,5 @@ export const recipeList = [
     ],
   }),
   Session.init(),
-  Dashboard.init({
-    apiKey: 'supertokens_is_awesome',
-  }),
+  Dashboard.init(),
 ];