+++ boilerplate/backend/python-fastapi/config/emailpassword.py
@@ -21,5 +21,5 @@ framework = "fastapi"
 recipe_list = [
     session.init(),
     emailpassword.init(),
-    dashboard.init(api_key="supertokens_is_awesome")
+    dashboard.init()
 ]