+++ boilerplate/fullstack/next/config/backend/thirdpartyemailpassword.ts
@@ -38,9 +38,7 @@ export let backendConfig = (): AuthConfig => {
                 ],
             }),
             SessionNode.init(),
-            Dashboard.init({
-                apiKey: "supertokens_is_awesome",
-            }),
+            Dashboard.init(),
         ],
         isInServerlessEnv: true,
     };