+++ boilerplate/fullstack/next/pages/auth/[[...path]].tsx
@@ -4,7 +4,10 @@ import styles from "../../styles/Home.module.css";
 import dynamic from "next/dynamic";
 import SuperTokens from "supertokens-auth-react";
 
-const SuperTokensComponentNoSSR = dynamic(new Promise((res) => res(SuperTokens.getRoutingComponent)), { ssr: false });
+const SuperTokensComponentNoSSR = dynamic<React.ComponentProps<typeof SuperTokens.getRoutingComponent>>(
+    new Promise((res) => res(SuperTokens.getRoutingComponent)),
+    { ssr: false }
+);
 
 export default function Auth(): JSX.Element {
     useEffect(() => {
@@ -19,6 +22,11 @@ export default function Auth(): JSX.Element {
         <div className={styles.container}>
             <Head>
                 <title>SuperTokens 💫</title>
+                <link
+                    href="//fonts.googleapis.com/css2?family=Rubik:wght@400&display=swap"
+                    rel="stylesheet"
+                    type="text/css"
+                />
                 <link rel="icon" href="/favicon.ico" />
             </Head>